requirements.txt
accelerate==1.0.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
anytree==2.12.1
appdirs==1.4.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
autograd==1.7.0
backcall==0.2.0
beartype==0.18.5
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bleach==6.1.0
blinker==1.9.0
bm25s==0.2.2
cachetools==5.5.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
evaluate==0.4.3
fastapi==0.115.8
fastapi-cli==0.0.7
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.23.5
hyppo==0.4.0
idna==3.10
ipython==8.12.3
Jinja2==3.1.5
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.134
llama-cloud==0.1.5
llama-index==0.12.2
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.2
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.0
llama-index-extractors-entity==0.3.0
llama-index-graph-stores-neo4j==0.4.0
llama-index-indices-managed-llama-cloud==0.6.2
llama-index-legacy==0.9.48.post4
llama-index-llms-huggingface==0.4.0
llama-index-llms-huggingface-api==0.3.0
llama-index-llms-ollama==0.4.0
llama-index-llms-openai==0.3.1
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.0
llama-index-vector-stores-postgres==0.3.0
llama-parse==0.5.7
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
minijinja==2.2.0
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.26.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.26.0
networkx==3.4
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
ollama==0.3.3
openai==1.51.2
orjson==3.10.7
pandas==2.2.3
pandocfilters==1.5.1
patsy==0.5.6
pgvector==0.2.5
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
POT==0.9.4
propcache==0.2.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
PyMuPDF==1.24.11
pynndescent==0.5.13
pyparsing==3.2.0
pypdf==5.1.0
pysbd==0.3.4
PyStemmer==2.2.0.3
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
ragas==0.2.7
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==3.2.0
seqeval==1.2.2
shellingham==1.5.4
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
span-marker==1.5.0
SQLAlchemy==2.0.35
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.42.0
striprtf==0.0.26
sympy==1.13.3
tenacity==8.5.0
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
wrapt==1.16.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.14.0
